Starting to add chemistry.
It is here that I realize more work is required. The paradigm
of mixture vs. (thread-safe) evaluator has not been adhered to.
Going to need to create a PlanetPhyscisEvaluator class that is
created at thread creation time and going to take a
PlanetPhysicsHelper class (which will be cached in the
PlanetPhysics class).